% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCor.R
\name{BBvar}
\alias{BBvar}
\title{Calculate block-to-block semivariogram}
\usage{
BBvar(model, lat = 0, rsl, dists = c(0, 1:7 * 10 + 2))
}
\arguments{
\item{model}{A variogram model object from the \code{gstat} package.}

\item{lat}{Latitude in degrees, defaults to 0.}

\item{rsl}{Resolution vector with x and y resolution in degrees (e.g., c(xres, yres)).}

\item{dists}{Vector of distances (in km) for which to calculate the semivariogram,
defaults to c(0, 12, 22, 32, 42, 52, 62, 72).}
}
\value{
An object of class \code{gstatVariogram} containing the block-to-block semivariogram.
}
\description{
Computes the block-to-block semivariogram based on a variogram model, discretising a block into pixels
and calculating semivariances within and between blocks at specified distances.
}
\examples{
  # Create a simple spherical variogram model and compute block semivariogram:
  library(gstat)
  vgm_model <- vgm(psill = 1, model = "Sph", range = 100, nugget = 0.1)
  res <- c(0.01, 0.01)  # 0.01-degree resolution
  result <- BBvar(model = vgm_model, lat = 40, rsl = res)
  print(result)
}
