% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCor.R
\name{blockCorr}
\alias{blockCorr}
\title{Compute block covariance matrix}
\usage{
blockCorr(lat, rsl, model, nx, ny)
}
\arguments{
\item{lat}{Latitude in degrees.}

\item{rsl}{Resolution vector with x and y resolution in degrees (e.g., c(xres, yres)).}

\item{model}{A variogram model object from the \code{gstat} package.}

\item{nx}{Number of pixels in the x-direction.}

\item{ny}{Number of pixels in the y-direction.}
}
\value{
A numeric matrix representing the covariance between all pairs of points within the block.
}
\description{
Calculates the covariance matrix for a block based on a variogram model, discretising the block into
pixels and computing covariances between all pairs of points within the block.
}
\examples{
 #Compute covariance matrix for a 5x5 pixel block:
  library(gstat)
  vgm_model <- vgm(psill = 1, model = "Sph", range = 100, nugget = 0.1)
  res <- c(0.02, 0.02)  # 0.02-degree resolution
  result <- blockCorr(lat = 40, rsl = res, model = vgm_model, nx = 5, ny = 5)
  dim(result)  # Should be 25 x 25
}
