% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecover.R
\name{calculate_gfc_tiles}
\alias{calculate_gfc_tiles}
\title{Calculate the GFC product tiles needed for a given area}
\usage{
calculate_gfc_tiles(aoi, recreate_grid = FALSE)
}
\arguments{
\item{aoi}{An sf object representing the Area of Interest}

\item{recreate_grid}{Logical. If TRUE, force recreation of the GFC tiles grid}
}
\value{
A filtered sf object containing only the tiles that intersect with the AOI
}
\description{
Determines which Global Forest Change (GFC) product tiles intersect
with a specified area of interest.
}
\examples{
\dontrun{
library(sf)

# Create a polygon in Central Africa
aoi <- st_sf(geometry = st_sfc(st_polygon(list(cbind(
  c(20, 21, 21, 20, 20),
  c(-1, -1, 1, 1, -1)))), crs = 4326))

tiles <- calculate_gfc_tiles(aoi)
}
}
