% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_and_convert_plt}
\alias{check_and_convert_plt}
\title{Check and covert plt input}
\usage{
check_and_convert_plt(plt, ez = FALSE)
}
\arguments{
\item{plt}{A data frame or sf object containing plot data. For data frame input format, longitude and latitude coordinates should be placed under "POINT_X" and "POINT_Y" columns respectively in WGS 84 CRS.}
}
\value{
A dataframe with "POINT_X" and "POINT_Y" columns
}
\description{
This helper function checks if the input is an sf object or a dataframe,
and ensures it contains the required coordinate columns.
}
\details{
If the input is an sf object, it is converted to a dataframe with coordinate columns.
The function then checks for the presence of "POINT_X" and "POINT_Y" columns.
}
