% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecover.R
\name{create_gfc_tiles_grid}
\alias{create_gfc_tiles_grid}
\title{Create a global grid of GFC tiles}
\usage{
create_gfc_tiles_grid()
}
\value{
An \code{sf} object representing the global grid of GFC tiles.
}
\description{
Creates a global grid of 10x10 degree tiles covering the world,
matching the Hansen Global Forest Change product naming convention.
}
\keyword{internal}
