% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deforested.R
\name{Deforested}
\alias{Deforested}
\title{Remove deforested plots by overlaying plots with Global Forest Change data}
\usage{
Deforested(plt, fdir, dataset = 2023, map_year = 10)
}
\arguments{
\item{plt}{A data frame or sf object containing plot data. Coordinates are assumed to be in WGS 84 CRS.}

\item{fdir}{Character string specifying the directory to download GFC forest loss data.}

\item{dataset}{Numeric value describing which version of the Hansen data to use: 2023 (default) down to 2018.}

\item{map_year}{Numeric value indicating the year threshold for deforestation (default is 10).}
}
\value{
A list containing two elements:
\item{netPlt}{A data frame of non-deforested plots}
\item{plt}{The original input data frame with an added deforestation indicator}
}
\description{
This function identifies and removes plots that have been deforested based on
the Global Forest Change dataset (Hansen et al., 2013). It processes each plot,
downloads the necessary forest loss tiles, and determines if the plot has been
deforested beyond a 5\% threshold or if the deforestation occurred before or during the specified map year.
}
\examples{
\dontrun{
  plot_data <- data.frame(POINT_X = c(-3.1, -3.2), POINT_Y = c(51.5, 51.6),
                          SIZE_HA = c(1, 1.5), PLOT_ID = c(1, 2))
  result <- Deforested(plot_data, "path/to/forest/loss/data", 10)
}
}
