% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefLidar.R
\name{detect_agb_layer}
\alias{detect_agb_layer}
\title{Detect and extract AGB layer from multi-band rasters}
\usage{
detect_agb_layer(raster_file)
}
\arguments{
\item{raster_file}{Path to raster file}
}
\value{
Single-layer terra raster with AGB data
}
\description{
Detect and extract AGB layer from multi-band rasters
}
