% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDIL4B.R
\name{download_gedi_l4b}
\alias{download_gedi_l4b}
\title{Download GEDI L4B Gridded Biomass Data}
\usage{
download_gedi_l4b(file_url, gedi_folder = "data/GEDI_L4B", max_tries = 3)
}
\arguments{
\item{file_url}{The URL of the GEDI L4B file to download.}

\item{max_tries}{The maximum number of retry attempts for the download.}

\item{output_path}{The local path where the file should be saved.}
}
\value{
A message indicating whether the download was successful or not.
}
\description{
This function downloads GEDI L4B Gridded Biomass data from the ORNL DAAC.
It uses the httr2 package for the HTTP request and handles authentication
using the earthdatalogin package if necessary.
}
