% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESACCIBiomass.R
\name{ESACCIAGBtileNames}
\alias{ESACCIAGBtileNames}
\title{Generate ESA-CCI AGB tile names}
\usage{
ESACCIAGBtileNames(
  pol,
  esacci_biomass_year = "latest",
  esacci_biomass_version = "latest"
)
}
\arguments{
\item{pol}{An sf or SpatVector object representing the polygon of interest.}

\item{esacci_biomass_year}{The ESACCI BIOMASS AGB tiles year to use. Use either 2010, 2015, 2016, 2017, 2018, 2019,
2020, 2021 or "latest" (default).}

\item{esacci_biomass_version}{The ESACCI BIOMASS AGB tiles version to use. Use either "v2.0", "v3.0", "v4.0",
"v5.0", "v5.01" or "latest" (default).}
}
\value{
A character vector of unique file names for ESA-CCI AGB tiles.
}
\description{
This function generates file names for ESA-CCI AGB tiles based on a given polygon.
}
