% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementErr.R
\name{growthUncertainty}
\alias{growthUncertainty}
\title{Calculate Growth Model Uncertainty}
\usage{
growthUncertainty(genus, species, growthParams)
}
\arguments{
\item{genus}{Character vector of tree genera}

\item{species}{Character vector of tree species}

\item{growthParams}{Data frame containing growth parameters by species}
}
\value{
Numeric vector of growth model standard deviations
}
\description{
Estimates uncertainty from biomass growth predictions using species-specific growth parameters
}
