% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementErr.R
\name{MeasurementErr}
\alias{MeasurementErr}
\title{Calculate Plot-Level AGB and Standard Deviation}
\usage{
MeasurementErr(plot, xy, region = "World")
}
\arguments{
\item{plot}{A data frame containing tree-level data.}

\item{xy}{A data frame containing plot locations.}
}
\value{
A data frame with plot-level AGB estimates and standard deviations, including columns:
PLOT_ID, POINT_X, POINT_Y, SIZE_HA, AVG_YEAR, AGB_T_HA, and sdTree.
}
\description{
This function calculates plot-level Above Ground Biomass (AGB) and Standard Deviation using tree-level data and plot locations.
It uses the BIOMASS package for AGB calculation and uncertainty estimation.
}
\details{
The function performs the following steps:
\enumerate{
\item Filters trees with diameter >= 10cm
\item Corrects taxonomy and retrieves wood density data
\item Computes or uses provided tree height data
\item Runs Monte Carlo simulation for AGB estimation
\item Calculates plot-level AGB and standard deviation
\item Scales values per hectare
}
}
\examples{
\dontrun{
plotsTree <- utils::read.csv(sample_file("SampleTree.csv"))
xyTree <- utils::read.csv(sample_file("SampleTreeXY.csv"))
plot_uncertainties <- MeasurementErr(plotsTree, xyTree, region = "Asia")
}
}
