% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plots}
\alias{plots}
\title{Sample plots data (dataframe points)}
\format{
\subsection{\code{plots}}{

A data frame with 8321 rows and 6 columns:
\describe{
\item{PLOT_ID}{Plot ID}
\item{POINT_X}{Longitude}
\item{POINT_Y}{Latitude}
\item{AGB_T_HA}{Above Ground Biomass (t/ha)}
\item{AVG_YEAR}{Average year of plot survey}
\item{SIZE_HA}{Plot size (ha)}
}
}
}
\usage{
plots
}
\description{
A  dataframe containing sample plots with longitude and latitude point data, Above Ground Biomass in t/ha,
average year of plot survey, and plot size in ha (SIZE_HA).
}
\keyword{datasets}
