% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygonize.R
\name{polyCirc}
\alias{polyCirc}
\title{Create circular plots}
\usage{
polyCirc(coords_poly, radius = 10)
}
\arguments{
\item{coords_poly}{A data frame with POINT_X and POINT_Y columns.}

\item{radius}{Numeric, the radius of the circular plot in meters (default: 10).}
}
\value{
An sf object representing the circular plot.
}
\description{
Create circular plots
}
\keyword{internal}
