% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygonize.R
\name{polyIrreg}
\alias{polyIrreg}
\title{Create irregular polygons}
\usage{
polyIrreg(coords_poly)
}
\arguments{
\item{coords_poly}{A data frame with POINT_X and POINT_Y columns.}
}
\value{
An sf object representing the irregular polygon.
}
\description{
Create irregular polygons
}
\keyword{internal}
