% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecover.R
\name{process_pair}
\alias{process_pair}
\title{Process a single pair of tree cover and forest loss files}
\usage{
process_pair(
  id,
  year,
  forest_loss_threshold,
  output_folder,
  treecover_files,
  lossyear_files,
  treecover_ids,
  lossyear_ids
)
}
\arguments{
\item{id}{The identifier for the file pair}

\item{year}{The year for which to compute tree cover}

\item{forest_loss_threshold}{The threshold for forest loss}

\item{output_folder}{The folder to save output rasters}

\item{treecover_files}{List of tree cover files}

\item{lossyear_files}{List of forest loss files}

\item{treecover_ids}{IDs for tree cover files}

\item{lossyear_ids}{IDs for forest loss files}
}
\value{
A SpatRaster object representing the computed tree cover
}
\description{
Process a single pair of tree cover and forest loss files
}
