% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sample_lidar_folder}
\alias{sample_lidar_folder}
\title{Get path to example sample lidar folder}
\usage{
sample_lidar_folder(folder)
}
\arguments{
\item{folder}{Name of folder within the extdata directory.}
}
\description{
This package comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
\dontrun{
sample_lidar_folder("TERN_data_package_v01/TERN_AGBmaps")
}
}
