% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockMeans.R
\name{sampleAGBmap}
\alias{sampleAGBmap}
\title{Sample block mean mapped AGB over a polygon}
\usage{
sampleAGBmap(pol, wghts = FALSE, own = TRUE, AGBown)
}
\arguments{
\item{pol}{An sf or SpatVector object representing the polygon of interest.}

\item{wghts}{Logical, if TRUE the weighted mean is calculated considering the
approximate fraction of each cell that is covered by the polygon (default is FALSE).}

\item{own}{Logical, whether to use a custom AGB map (default is TRUE).}

\item{AGBown}{A SpatRaster object with the custom AGB map.}
}
\value{
Numeric value representing the mean AGB for the polygon.
}
\description{
This function samples the block mean mapped Above Ground Biomass (AGB) over a given polygon.
It can use either a custom AGB map or predefined AGB tiles.
}
