% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockMeans.R
\name{sampleTreeCover}
\alias{sampleTreeCover}
\title{Sample block mean mapped forest over a polygon}
\usage{
sampleTreeCover(roi, thresholds, weighted_mean = FALSE, fmask = NA)
}
\arguments{
\item{roi}{An sf or SpatVector object representing the Region of Interest.}

\item{thresholds}{Numeric vector of tree cover thresholds to apply.}

\item{weighted_mean}{Logical, if TRUE the weighted mean is calculated considering the
approximate fraction of each cell that is covered by the polygon (default is FALSE).}

\item{fmask}{A SpatRaster object representing a custom forest mask (default is NA).}
}
\value{
Numeric vector of tree cover percentages for each threshold.
}
\description{
This function samples the block mean mapped forest cover over a given polygon.
It can use either predefined tree cover tiles or a custom forest mask.
}
