% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementErr.R
\name{samplingUncertainty}
\alias{samplingUncertainty}
\title{Calculate Sampling Error Uncertainty}
\usage{
samplingUncertainty(plotAGB, plotSize, method = "basic")
}
\arguments{
\item{plotAGB}{Numeric vector of plot-level AGB values}

\item{plotSize}{Numeric vector of plot sizes in hectares}

\item{method}{Character specifying error propagation method ("basic" or "bootstrap")}
}
\value{
Numeric vector of sampling error standard deviations
}
\description{
Computes standard error of the mean for plot biomass estimates using standard error propagation methods
}
