% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileNames.R
\name{SDtileNames}
\alias{SDtileNames}
\title{Generate AGB SD tile names}
\usage{
SDtileNames(
  pol,
  agbTilesDir = "data/ESACCI-BIOMASS-L4-AGB-MERGED-100m-2020-fv5.0"
)
}
\arguments{
\item{pol}{An sf or SpatVector object representing the polygon of interest.}
}
\value{
A character vector of unique file names for AGB SD tiles.
}
\description{
This function generates file names for AGB SD tiles based on a given polygon.
}
