% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sf_to_df_with_coords}
\alias{sf_to_df_with_coords}
\title{Convert sf object to dataframe with longitude and latitude columns}
\usage{
sf_to_df_with_coords(sf_object)
}
\arguments{
\item{sf_object}{An sf object with POINT geometry}
}
\value{
A data frame with all original columns plus POINT_X and POINT_Y in WGS 84 CRS.
}
\description{
This function takes an sf object and converts it to a data frame,
adding "POINT_X" and "POINT_Y" columns for longitude and latitude.
If the input sf object is not in WGS 84, it will be transformed.
}
\examples{
library(sf)
# Assuming 'sampled_plots$non_deforested_plots' is your sf object
df <- sf_to_df_with_coords(sampled_plots$non_deforested_plots)
}
