% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefLidar.R
\name{standardize_year}
\alias{standardize_year}
\title{Standardize year format (handle 2-digit years)}
\usage{
standardize_year(year_str)
}
\arguments{
\item{year_str}{Character vector of year strings}
}
\value{
Numeric vector of standardized years
}
\description{
Standardize year format (handle 2-digit years)
}
