% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileNames.R
\name{TCtileNames}
\alias{TCtileNames}
\title{Generate tree cover tile names}
\usage{
TCtileNames(pol, year = 2010, treeCoverDir = "data/treecover2010_v3_100m")
}
\arguments{
\item{pol}{An sf or SpatVector object representing the polygon of interest.}

\item{year}{Numeric, the year for which to generate tile names (2010, 2015, or 2020).}
}
\value{
A character vector of unique file names for tree cover tiles.
}
\description{
This function generates file names for tree cover tiles based on a given polygon and year.
}
