% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempApplyVar.R
\name{TempApplyVar}
\alias{TempApplyVar}
\title{Apply temporal adjustment and compute temporal variance in plot biomass}
\usage{
TempApplyVar(plt, map_year, gez = "all", gr = NULL, sds = NULL)
}
\arguments{
\item{plt}{A data frame or sf object containing plot data. For data frame input format, longitude and latitude coordinates should be placed under "POINT_X" and "POINT_Y" columns respectively in WGS 84 CRS.}

\item{map_year}{Numeric value indicating the AGB map year.}

\item{gez}{Character string specifying the Global Ecological Zones (GEZ) of interest to apply. If "all" (default),
all GEZ in plt will be calculated.}

\item{gr}{Optional data.frame of growth rates containing columns \code{GEZ}, \code{ZONE},
\code{FAO.ecozone}, \code{GR1}, \code{GR2}, \code{GR3}. If \code{NULL}, default data
from \pkg{Plot2Map} is used.}

\item{sds}{Optional data.frame of growth rate standard deviations containing columns
\code{GEZ}, \code{ZONE}, \code{FAO.ecozone}, \code{SD1}, \code{SD2}, \code{SD3}. If
\code{NULL}, default data from \pkg{Plot2Map} is used.}
}
\value{
A data.frame with the following columns:
\describe{
\item{\code{AGB_T_HA}}{Adjusted biomass values for the specified \code{map_year}.}
\item{\code{AGB_T_HA_ORIG}}{Original biomass values before adjustment.}
\item{\code{sdGrowth}}{Temporal standard deviation of the biomass adjustment.}
\item{All other columns from the input \code{plt}.}
}
}
\description{
This function adjusts plot biomass values to a specified map year using growth-rate increments
and computes the temporal variance (standard deviation) of biomass changes based on growth-rate
standard deviations. It combines the functionality of \code{\link{TempApply}} and
\code{\link{TempVar}} into a single call.
}
\examples{
  library(Plot2Map)
  set.seed(42)
  sample_plots <- plots[sample(nrow(plots), 10), ]
  sample_plots_gez <- BiomePair(sample_plots)
  result <- TempApplyVar(sample_plots_gez, map_year = 2004)
  head(result)
}
