% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempFix.R
\name{TempVar}
\alias{TempVar}
\title{Calculate temporal variance in plot biomass}
\usage{
TempVar(df, domain, year)
}
\arguments{
\item{df}{A data frame containing plot data.}

\item{domain}{Character string specifying the ecological domain of interest.}

\item{year}{Numeric value indicating the map AGB year (e.g., 2010 for GlobBiomass).}
}
\value{
A data frame with adjusted AGB values and calculated growth standard deviations.
}
\description{
This function calculates the temporal variance in plot biomass based on growth rate
standard deviations. It adjusts the AGB values and computes the standard deviation
of growth for each plot.
}
