% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESACCIBiomass.R
\name{validate_esacci_biomass_args}
\alias{validate_esacci_biomass_args}
\title{Helper function used for args validation}
\usage{
validate_esacci_biomass_args(esacci_biomass_year, esacci_biomass_version)
}
\description{
Helper function used for args validation
}
