% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecover.R
\name{compute_treecover}
\alias{compute_treecover}
\title{Compute Tree Cover for a Specific Year}
\usage{
compute_treecover(
  year,
  gfc_folder = "data/GFC",
  forest_loss_threshold = NULL,
  coords = NULL,
  output_folder = NULL,
  num_cores = 1
)
}
\arguments{
\item{year}{Integer. The year for which to compute the tree cover (2001-2023).}

\item{gfc_folder}{Character. Path to the folder containing GFC dataset files. Default is "data/GFC".}

\item{forest_loss_threshold}{Numeric. Optional threshold for forest loss (0-100). If provided, only areas with
forest loss values >= this threshold will be considered as loss. Default is NULL (no threshold).}

\item{coords}{Character. Coordinates of the tile to process, e.g., "10N_020E". If NULL, processes all available tiles.}

\item{output_folder}{Character. Path to the folder where output rasters should be saved. If NULL, rasters are not saved to disk. Default is NULL.}

\item{num_cores}{Integer. Number of cores to use for parallel processing.}
}
\value{
If coords is specified or only one tile is processed, returns a single SpatRaster object.
Otherwise, returns a list of SpatRaster objects, each representing tree cover for the specified year for different tiles.
}
\description{
This function computes a tree cover layer for a given year based on the Hansen Global Forest Change (GFC) dataset.
It uses the tree cover from 2000 as a baseline and subtracts forest loss up to the specified year.
}
\details{
The function uses two pattern matched (matched by GFC dataset year, version and lat, log coordinates)
types of input files stored in the specified gfc_folder, for example:
\enumerate{
\item Tree cover 2000: "Hansen_GFC-2023-v1.11_treecover2000_*.tif"
\itemize{
\item Values represent percentage of tree cover (0-100\%)
}
\item Forest loss year: "Hansen_GFC-2023-v1.11_lossyear_*.tif"
\itemize{
\item Values: 0 (no loss) or 1-23 (loss year 2001-2023)
}
}
}
\examples{
\dontrun{
# Compute tree cover for 2015 for all available tiles
treecover_2015 <- compute_treecover(2015)

# Compute tree cover for 2020 for a specific tile with a forest loss threshold of 30\% and save output
treecover_2020 <- compute_treecover(2020,
                                    forest_loss_threshold = 30,
                                    coords = "10N_020E",
                                    output_folder = "data/GFC_calc")
}
}
