% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempVis.R
\name{HistoShift}
\alias{HistoShift}
\title{Analyze AGB shift in bins}
\usage{
HistoShift(df, year, outDir = "results")
}
\arguments{
\item{df}{A data frame containing AGB data.}

\item{year}{Numeric value indicating the year of analysis.}

\item{outDir}{Character string specifying the output directory for saving the CSV file (default: "results").}
}
\value{
A data frame summarizing the changes in AGB distribution across bins.
}
\description{
This function analyzes the shift in AGB values across different bins before and after
temporal adjustment or forest fraction correction. It generates a summary table
of the changes in AGB distribution.
}
\examples{
set.seed(42)
sample_plots <- plots[sample(nrow(plots), 10), ]
sample_plots <- BiomePair(sample_plots)
sample_plots <- TempApply(sample_plots, 2004)
head(sample_plots)
HistoShift(sample_plots, 2004)
}
