% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBlockPolygon.R
\name{MakeBlockPolygon}
\alias{MakeBlockPolygon}
\title{Create a block polygon}
\usage{
MakeBlockPolygon(x, y, size)
}
\arguments{
\item{x}{Numeric, x-coordinate of a point within the desired block.}

\item{y}{Numeric, y-coordinate of a point within the desired block.}

\item{size}{Numeric, the size of the square block.}
}
\value{
An sf object representing the square polygon.
}
\description{
This function creates a square polygon (block) based on given coordinates and size.
It calculates the lower left corner of the cell and defines the vertices of the polygon.
}
\examples{
block <- MakeBlockPolygon(5.2, 10.7, 1)
}
