% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested.R
\name{Nested}
\alias{Nested}
\title{Format tree-level data from nested plots}
\usage{
Nested(centroids_sf, tree_table)
}
\arguments{
\item{centroids_sf}{A sf object containing centroids information of sub-plots.}

\item{tree_table}{A data frame containing tree-level measurements.}
}
\value{
A list containing two elements:
\item{plotTree}{A data frame with formatted tree-level data}
\item{xyTree}{A data frame with plot ID and coordinates}
}
\description{
This function processes tree-level data from nested plots, combining centroid information
with tree measurements using a unique ID from these datasets.
}
\details{
The function performs the following steps:
\enumerate{
\item Converts centroid data to WGS84 projection
\item Creates a buffer around centroids
\item Filters tree data for alive trees
\item Prompts user for column selections and additional information
\item Processes tree height data
\item Combines tree data with centroid coordinates
}
}
\examples{
centroid_data <- readOGR(dsn = "path/to/shapefile", layer = "centroid_layer")
tree_data <- read.csv("path/to/tree_data.csv")
nested_data <- Nested(centroid_data, tree_data)
pol
}
