% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecover.R
\name{process_pair}
\alias{process_pair}
\title{Process a single pair of tree cover and forest loss files}
\usage{
process_pair(
  id,
  year,
  treecover_threshold,
  output_folder,
  treecover_files,
  lossyear_files,
  treecover_ids,
  lossyear_ids,
  baseline
)
}
\arguments{
\item{id}{The identifier for the file pair}

\item{year}{The year for which to compute tree cover}

\item{treecover_threshold}{The threshold for tree cover}

\item{output_folder}{The folder to save output rasters}

\item{treecover_files}{List of tree cover files}

\item{lossyear_files}{List of forest loss files}

\item{treecover_ids}{IDs for tree cover files}

\item{lossyear_ids}{IDs for forest loss files}

\item{baseline}{Baseline year for tree cover dataset (2000 or 2010)}
}
\value{
A file path to a geotiff raster representing the computed tree cover
}
\description{
Process a single pair of tree cover and forest loss files
}
