% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockMeans.R
\name{sampleAGBmap}
\alias{sampleAGBmap}
\title{#' Sample block mean mapped AGB over a region of interest
#'
#' This function samples the block mean mapped Above Ground Biomass (AGB) over a given polygon.
#' It can use either a custom AGB map or predefined AGB tiles.
#'
#' @param roi An sf or SpatVector object representing the Region of Interest.
#' @param weighted_mean Logical, if TRUE the weighted mean is calculated considering the
#'  approximate fraction of each cell that is covered by the polygon (default is FALSE).
#' @param agb_raster A SpatRaster object with the custom AGB map. If NULL, ESACCI BIOMASS AGB tiles will be used.
#' This dataset comprises estimates of forest above-ground biomass for the years 2010, 2015, 2016, 2017, 2018, 2019,
#' 2020 and 2021. They are derived from a combination of Earth observation data, depending on the year, from the Copernicus
#' Sentinel-1 mission, Envisat’s ASAR (Advanced Synthetic Aperture Radar) instrument and JAXA’s (Japan Aerospace Exploration
#' Agency) Advanced Land Observing Satellite (ALOS-1 and ALOS-2), along with additional information from Earth observation
#' sources. The data has been produced as part of the European Space Agency's (ESA's) Climate Change Initiative (CCI)
#' programme by the Biomass CCI team.
#' the given year.
#' @param esacci_biomass_year The ESACCI BIOMASS AGB tiles year to use when 'agb_raster = NULL'. Use either 2010, 2015, 2016, 2017, 2018, 2019,
#' 2020, 2021 or "latest" (default).
#' @param esacci_biomass_version The ESACCI BIOMASS AGB tiles version to use when 'agb_raster = NULL'. Use either "v2.0", "v3.0", "v4.0",
"v5.0", "v5.01" or "latest" (default).
#'
#' @return Numeric value representing the mean AGB for the polygon.
#'
#' @importFrom sf st_as_sf
#' @importFrom terra rast extract
#'
#' @export
sampleAGBmap <- function(roi, weighted_mean = FALSE, agb_raster = NULL, esacci_biomass_year = "latest") {
Sample block mean mapped AGB over a polygon}
\usage{
sampleAGBmap(pol, wghts = FALSE, own = TRUE, AGBown)
}
\arguments{
\item{pol}{An sf or SpatVector object representing the polygon of interest.}

\item{wghts}{Logical, if TRUE the weighted mean is calculated considering the
approximate fraction of each cell that is covered by the polygon (default is FALSE).}

\item{own}{Logical, whether to use a custom AGB map (default is TRUE).}

\item{AGBown}{A SpatRaster object with the custom AGB map.}
}
\value{
Numeric value representing the mean AGB for the polygon.
}
\description{
This function samples the block mean mapped Above Ground Biomass (AGB) over a given polygon.
It can use either a custom AGB map or predefined AGB tiles.
}
