% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockMeans.R
\name{sampleTreeCover}
\alias{sampleTreeCover}
\title{Sample block mean mapped forest over a polygon}
\usage{
sampleTreeCover(pol, thresholds, wghts = FALSE, fmask = NA)
}
\arguments{
\item{pol}{An sf or SpatVector object representing the polygon of interest.}

\item{thresholds}{Numeric vector of tree cover thresholds to apply.}

\item{wghts}{Logical, whether to use weighted extraction (default is FALSE).}

\item{fmask}{A SpatRaster object representing a custom forest mask (default is NA).}
}
\value{
Numeric vector of tree cover percentages for each threshold.
}
\description{
This function samples the block mean mapped forest cover over a given polygon.
It can use either predefined tree cover tiles or a custom forest mask.
}
