% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempFix.R
\name{TempApply}
\alias{TempApply}
\title{Apply temporal adjustment to plot biomass}
\usage{
TempApply(plt, map_year, gez = "all")
}
\arguments{
\item{plt}{A data frame or sf object containing plot data. For data frame input format, longitude and latitude coordinates should be placed under "POINT_X" and "POINT_Y" columns respectively in WGS 84 CRS.}

\item{map_year}{Numeric value indicating the AGB map year.}

\item{gez}{Character string specifying the Global Ecological Zones (GEZ) of interest to apply. If "all" (default),
all GEZ in plt will be calculated.}
}
\value{
A data frame with adjusted AGB values.
}
\description{
This function adjusts plot biomass values to align with the map year by adding or subtracting
annual growth increment to older or newer plots. It uses growth data from a model-based
estimate of growth-age relationships.
}
