% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempFix.R
\name{TempVar}
\alias{TempVar}
\title{Calculate temporal variance in plot biomass}
\usage{
TempVar(plt, map_year, gez = "all")
}
\arguments{
\item{plt}{A data frame or sf object containing plot data. For data frame input format, longitude and latitude coordinates should be placed under "POINT_X" and "POINT_Y" columns respectively in WGS 84 CRS.}

\item{map_year}{Numeric value indicating the AGB map year.}

\item{gez}{Character string specifying the Global Ecological Zones (GEZ) of interest to apply. If "all" (default),
all GEZ in plt will be calculated.}
}
\value{
A data frame with calculated growth standard deviations.
}
\description{
This function calculates the temporal variance in plot biomass based on growth rate
standard deviations. It computes the standard deviation
of growth for each plot.
}
